+++ app/assets/javascripts/answers.js
@@ -1,18 +1,5 @@
 $(document).on('turbolinks:load', function () {
   eventForEditAnswer();
-
-  // $('form.new-answer').on('ajax:success', function(e) {
-  //   var answer = e.detail[0];
-
-  //   $('.answers').append('<p>' + answer.body + '</p>');
-  // })
-  //   .on('ajax:error', function(e) {
-  //     var errors = e.detail[0];
-
-  //     $.each(errors, function(index, value) {
-  //       $('.create-answer-errors').append('<p>' + value + '</p>');
-  //     });
-  //   });
 });
 
 function eventForEditAnswer() {
