+++ app/controllers/answers_controller.rb
@@ -8,16 +8,6 @@ class AnswersController < ApplicationController
     @answer = @question.answers.new(answer_params)
     @answer.user_id = current_user.id
     @answer.save
-
-    # respond_to do |format|
-    #   if @answer.save
-    #     format.json { render json: @answer }
-    #   else
-    #     format.json do
-    #       render json: @answer.errors.full_messages, status: :unprocessable_entity
-    #     end
-    #   end
-    # end
   end
 
   def edit; end
