+++ app/models/answer.rb
@@ -1,13 +1,11 @@
 class Answer < ApplicationRecord
   include Linkable
-
-  after_save :create_score
+  include Scorable
 
   belongs_to :question
   belongs_to :user
 
   has_many_attached :files
-  has_one :score, dependent: :destroy, as: :scorable
 
   validates :body, presence: true
 
@@ -19,11 +17,5 @@ class Answer < ApplicationRecord
       update!(correct: true)
       user.medals.push(question.medal) if question.medal
     end
-  end
-
-  private
-
-  def create_score
-    Score.create!(author: user, scorable_type: "Answer", scorable_id: id)
   end
 end
