+++ spec/models/question_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 RSpec.describe Question, type: :model do
   include_examples "links"
-  it { should have_one(:score).dependent(:destroy) }
+  include_examples "scores", :question, 1
 
   it { should belong_to :user }
   it { should have_many(:answers).dependent(:destroy) }
@@ -16,10 +16,4 @@ RSpec.describe Question, type: :model do
 
   it { should validate_presence_of :title }
   it { should validate_presence_of :body }
-
-  describe "#create_score" do
-    it 'has to create score after creating itself' do
-      expect{ create(:question) }.to change(Score, :count).by(1)
-    end
-  end
 end
